fix: use uint64_t instead unsigned long long

+++ sysmonTypes.hpp
@@ -2,6 +2,7 @@
 #define SYSMON_TYPES_HPP
 
 #include <base/Time.hpp>
+#include <cstdint>
 #include <string>
 #include <vector>
 
@@ -11,22 +12,22 @@ namespace sysmon {
         unsigned int major_number = 0;
         unsigned int minor_number = 0;
         std::string device_name;
-        unsigned long long reads_completed = 0;
-        unsigned long long reads_merged = 0;
-        unsigned long long sectors_read = 0;
+        uint64_t reads_completed = 0;
+        uint64_t reads_merged = 0;
+        uint64_t sectors_read = 0;
         base::Time time_reading = base::Time();
-        unsigned long long writes_completed = 0;
-        unsigned long long writes_merged = 0;
-        unsigned long long sectors_written = 0;
+        uint64_t writes_completed = 0;
+        uint64_t writes_merged = 0;
+        uint64_t sectors_written = 0;
         base::Time time_writing = base::Time();
-        unsigned long long io_in_progress = 0;
+        uint64_t io_in_progress = 0;
         base::Time time_doing_io = base::Time();
         base::Time weighted_time_doing_io = base::Time();
-        unsigned long long discards_completed = 0;
-        unsigned long long discards_merged = 0;
-        unsigned long long sectors_discarded = 0;
+        uint64_t discards_completed = 0;
+        uint64_t discards_merged = 0;
+        uint64_t sectors_discarded = 0;
         base::Time time_discarding = base::Time();
-        unsigned long long flush_requests_completed = 0;
+        uint64_t flush_requests_completed = 0;
         base::Time time_flushing = base::Time();
     };
